Make double-doubles in fma big-endian like log family
Big-endianness is better when using double[2] for double-double
arithmetic.  Different from integer arithmetic, in a double-double
pair, the high part not only stores more significant bits but is
also the nearest representation of the whole pair.  Therefore, given

    double v[2];

It makes more sense that *v evaluates to the high part.  Moreover,
the high part is often accessed before the low part due to the
nature of floating-point arithmetic, contrary to integer arithmetic
again.